Add a second promo text to the front.
NRSS does not have enough supporters to break even and thus costs me
quite a bit every month. However, it seems to have a stable and high
user base, most of whom presumably could pay. As earlier, I'd like to
avoid shoving the service behind a paywall - that feels totally wrong
for what this project is supposed to be. This PR introduces another,
shorter link to donations _before_ the episode description in the
hopes of attracting more paid users. Let's see if it works :)

The last time there was a promo text at the start of the description,
@augustskare correctly pointed out that this hides the original description
from many views in podcast players in #34. That's not nice. I've tried to
acommodate this by making the initial description much shorter than
the one I had then (which now comes at the end).